backend/requirements.txt
Django>=4.2
djangorestframework>=3.14
django-cors-headers>=4.3
djangorestframework-simplejwt>=5.3.1
gunicorn>=23.0
whitenoise>=6.6
google-generativeai
